+++ CHANGELOG.md
@@ -17,10 +17,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+### Security
+
+## [v1.61.1] - 2022-07-14
+
+### Fixed
+
 - Revert not_before and not_after in jwt from time.Time to string
 - JWT test files format to match actual format from vault
-
-### Security
 
 ## [v1.61.0] - 2022-07-14
 
@@ -801,7 +805,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial state: standards for Main, HTTP Server, DB access (gorm), Logging, Instrumented-Handler, and K8s Probe
 
-[Unreleased]: https://github.com/gesundheitscloud/go-svc/compare/v1.61.0...HEAD
+[Unreleased]: https://github.com/gesundheitscloud/go-svc/compare/v1.61.1...HEAD
+[v1.61.1]: https://github.com/gesundheitscloud/go-svc/compare/v1.61.0...v1.61.1
 [v1.61.0]: https://github.com/gesundheitscloud/go-svc/compare/v1.60.0...v1.61.0
 [v1.60.0]: https://github.com/gesundheitscloud/go-svc/compare/v1.59.0...v1.60.0
 [v1.59.0]: https://github.com/gesundheitscloud/go-svc/compare/v1.58.0...v1.59.0